refactor: reorgnize file structure
