+++ cli/src/task/mod.rs
@@ -1,8 +1,14 @@
+mod static_lib;
 mod task;
 mod task_builder;
+mod wasm;
+mod xcframework;
 
 use super::function::*;
 
 pub(crate) use task::Task;
 
+pub use static_lib::*;
 pub use task_builder::TaskBuilder;
+pub use wasm::*;
+pub use xcframework::*;
